+++ packages/alfa-rules/test/sia-r62/rule.spec.tsx
@@ -928,3 +928,12 @@ test(`evaluate() passes an applicable <a> element that removes the default text
     }),
   ]);
 });
+
+test(`evaluate() is inapplicable to an <a> element with a <p> parent element
+    no non-link whitespace text content`, async (t) => {
+  const target = <a href="#">Link</a>;
+
+  const document = h.document([<p> {target}</p>]);
+
+  t.deepEqual(await evaluate(R62, { document }), [inapplicable(R62)]);
+});